feat: use pg-mem by default. Only use real postgres if DATABASE_URL set.
