+++ package.json
@@ -28,6 +28,7 @@
     "classnames": "^2.3.1",
     "mocha": "^9.2.1",
     "nanoid": "^3.3.1",
+    "pg-mem": "^2.5.0",
     "prettier": "^2.2.1",
     "pusher": "^5.1.0-beta",
     "pusher-js": "^7.1.1-beta",
